Atualização Bee-Api 240725

+++ src/controllers/controller.category.js
@@ -1,11 +1,12 @@
 import categoryService from "../services/service.category.js";
 
 const createCategoryController = async (req, res) => {
-  const { name, company_id } = req.body;
+  const { name, company_id, notification } = req.body;
   try {
     const category = await categoryService.createCategoryService(
       name,
-      company_id
+      company_id,
+      notification
     );
     return res.status(201).json(category);
   } catch (error) {
@@ -64,12 +65,13 @@ const getCategoriesByCompanyIdController = async (req, res) => {
 
 const updateCategoryController = async (req, res) => {
   const { category_id } = req.params;
-  const { name, company_id } = req.body;
+  const { name, company_id, notification } = req.body;
   try {
     const category = await categoryService.updateCategoryService(
       category_id,
       name,
-      company_id
+      company_id,
+      notification
     );
     if (!category)
       return res.status(404).json({ message: "Categoria não encontrada" });
@@ -81,8 +83,50 @@ const updateCategoryController = async (req, res) => {
 };
 
 const deleteCategoryController = async (req, res) => {
-  const { category_id } = req.params;
-  const { company_id } = req.query;
+  const category_id = parseInt(req.params.category_id, 10);
+  const company_id = parseInt(req.query.company_id, 10);
+
+  if (isNaN(category_id) || isNaN(company_id)) {
+    return res
+      .status(400)
+      .json({ message: "category_id e company_id devem ser números válidos" });
+  }
+
+  try {
+    // Chama service que deve fazer a verificação interna
+    const result = await categoryService.deleteCategoryService(
+      category_id,
+      company_id
+    );
+
+    if (result === "has_products") {
+      return res.status(400).json({
+        message:
+          "Não é possível deletar a categoria porque há produtos vinculados a ela.",
+      });
+    }
+
+    if (!result) {
+      return res.status(404).json({ message: "Categoria não encontrada" });
+    }
+
+    return res.status(200).json({ message: "Categoria deletada com sucesso" });
+  } catch (error) {
+    console.error("Erro ao deletar categoria: ", error);
+    return res.status(500).json({ message: "Erro ao deletar categoria" });
+  }
+};
+
+/*const deleteCategoryController = async (req, res) => {
+  const category_id = parseInt(req.params.category_id, 10);
+  const company_id = parseInt(req.query.company_id, 10);
+
+  if (isNaN(category_id) || isNaN(company_id)) {
+    return res
+      .status(400)
+      .json({ message: "category_id e company_id devem ser números válidos" });
+  }
+
   try {
     const category = await categoryService.deleteCategoryService(
       category_id,
@@ -95,7 +139,7 @@ const deleteCategoryController = async (req, res) => {
     console.error("Erro ao deletar categoria: ", error);
     return res.status(500).json({ message: "Erro ao deletar categoria" });
   }
-};
+};*/
 
 export default {
   createCategoryController,
